+++ CHANGELOG.md
@@ -2,9 +2,17 @@
 
 ## [unreleased]
 
+### Documentation
+
+- Add git-cliff config file and CHANGELOG.md
+
 ### Features
 
 - Add cargo boilerplate code
 - Add first prototype code
 - Implement first functional version
+
+### Refactor
+
+- Extract functions into own file functions.rs